Extract config to a separate module
